@@ -48,6 +48,13 @@
     "protractor": "~5.1.2",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "typescript": "^2.6.0",
+    "webpack": "^3.6.0",
+    "webpack-dev-server": "^2.9.1",
+    "angular2-template-loader": "^0.6.2",
+    "awesome-typescript-loader":"^3.3.0",
+    "uglifyjs-webpack-plugin":"^1.0.1",
+    "html-webpack-plugin": "^2.30.1",
+    "html-loader":"^0.5.1"
   }
 }
