@@ -11,12 +11,13 @@ import { RegInComponent } from './components/reg-in-component/reg-in.component';
 
 import {DashboardModule} from './dashboard-module/dashboard.module';
 import { NotFoundComponent } from './components/not-found-component/not-found.component';
-
+import { ManagerComponent } from './dashboard-module/manager-component/manager.component';
+import { DashboardComponent } from './dashboard-module/dashboard-component/dashboard.component';
+import { SideBarComponent } from './dashboard-module/side-bar-component/side-bar.component';
 const appRoutes: Routes =[
   { path: '', component: NotFoundComponent},
   { path: 'login', component: LoginComponent},
-  { path: 'regin', component: RegInComponent},
-  { path: '**', component: NotFoundComponent }
+  { path: 'regin', component: RegInComponent} //{ path: '**', component: NotFoundComponent },
 ];
 
 @NgModule({
