@@ -14,40 +14,23 @@ import {SideBarComponent} from '../side-bar-component/side-bar.component';
 export class DashboardComponent implements OnInit {
 public managers: Manager[]
 public manager: Manager
-  @Output()
-  public carList : Array<Car> = [
-    {name:"Mazda", stars: 1,id: 1},
-    {name:"Honda", stars: 2,id: 2},
-    {name:"Acura", stars: 1,id: 3},
-    {name:"Totota", stars: 4,id: 4},
-    {name:"Infinity", stars: 5,id: 5},
-    {name:"Infinity", stars: 5,id: 5},
-    {name:"Infinity", stars: 5,id: 5},
-    {name:"Infinity", stars: 5,id: 5},
-    {name:"Infinity", stars: 5,id: 5},
-    {name:"Infinity", stars: 5,id: 5},
-    {name:"Infinity", stars: 5,id: 5},
-    {name:"Infinity", stars: 5,id: 5},
-    {name:"Infinity", stars: 5,id: 5},
-    {name:"Infinity", stars: 5,id: 5},
-    {name:"Infinity", stars: 5,id: 5},
-  ]
+  
   constructor(private managerService: ManagerService) { 
   }
   ngOnInit() {
-    this.managerService.getManager(1)
-    .subscribe(manager => this.manager = manager)
-    this.getMangers()
-    debugger;
-    console.log("1 "+this.manager);
-    console.table("many "+this.managers);
+    //this.managerService.getManager(1)
+    //.subscribe(manager => this.manager = manager)
+    //this.getMangers()
+    //debugger;
+    //console.log("1 "+this.manager);
+    //console.table("many "+this.managers);
   }
 
   getMangers():void{
-    this.managerService.getManagers()
-    .subscribe(managers => {this.managers = managers
-    console.log(this.managers);
-    });
+    //his.managerService.getManagers()
+    //.subscribe(managers => {this.managers = managers
+    //console.log(this.managers);
+    //});
 
   }
   
