@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { HttpClient } from '@angular/common/http';
+import { HttpClient, HttpParams } from '@angular/common/http';
 import { Manager } from '../models/manager';
 import { Observable } from 'rxjs/Observable';
 import { of } from 'rxjs/observable/of';
@@ -8,9 +8,7 @@ import 'rxjs/add/operator/map';
 
 @Injectable()
 export class ManagerService {
-  constructor(private http: HttpClient) {
-
-  }
+  constructor(private http: HttpClient) { }
   public baseUrl: string = "http://localhost:14980/api" //14980
   public serviceUrl: string = "/managers"
 
@@ -19,17 +17,32 @@ export class ManagerService {
   //   return managers;
   // }
 
-  public getManager(id:number) {
+  public getManager(id: number) {
     let manager = this.http.get(this.baseUrl + this.serviceUrl + '/' + id);
     return manager;
   }
 
-  private extractManager(response: Response){
+  private extractManager(response: Response) {
     let json = response.json()
   }
-  getManagers (): Observable<Manager[]> {
+  getManagers(): Observable<Manager[]> {
     return this.http.get<Manager[]>(this.baseUrl + this.serviceUrl)
       .pipe(
       );
+  }
+  //getManagers() {
+    //return this.http.get(this.baseUrl);
+  //}
+
+  createManager(manager: Manager) {
+    return this.http.post(this.baseUrl, manager);
+  }
+  updateManager(id: number, manager: Manager) {
+    const urlParams = new HttpParams().set("id", id.toString());
+    return this.http.put(this.baseUrl, manager, { params: urlParams });
+  }
+  deleteManager(id: number) {
+    const urlParams = new HttpParams().set("id", id.toString());
+    return this.http.delete(this.baseUrl, { params: urlParams });
   }
 }
